TalkativeAI/speech_service/requirements.txt
fastapi
uvicorn
openai-whisper
torch
ffmpeg-python
pydantic
